requirements.txt
amqp==1.4.6
anyjson==0.3.3
BeautifulSoup==3.2.1
billiard==3.3.0.20
celery==3.1.18
Django==1.7.5
django-celery==3.1.16
django-common-helpers==0.7.0
django-filter==0.10.0
django-mysql==0.1.8
django-rest-auth==0.4.0
djangorestframework==3.1.3
djangotoolbox==0.9.2
djangular==0.2.7
enum34==1.0.4
kombu==3.0.26
Markdown==2.6.2
mechanize==0.2.5
MySQL-python==1.2.5
parse==1.6.6
parse-type==0.3.4
pytz==2015.4
requests==2.8.1
six==1.9.0
wsgi-intercept==0.5.1
XlsxWriter==0.7.3
